Suppress compiler warnings
streams.cc: In function ‘void grow(Stream*, int, int)’:
streams.cc:65:15: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  if (newlen > stream_alloc_maximum) {
               ^
streams.cc:66:30: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
      if (s->current + need < stream_alloc_maximum)

+++ streams.cc
@@ -62,8 +62,8 @@ grow(Stream * s, int newlen, int need)
     char *newbuf;
 
     if (allow_stream_exceptions > 0) {
-	if (newlen > stream_alloc_maximum) {
-	    if (s->current + need < stream_alloc_maximum)
+	if (newlen > (int)stream_alloc_maximum) {
+	    if (s->current + need < (int)stream_alloc_maximum)
 		newlen = stream_alloc_maximum;
 	    else
 		throw stream_too_big();
